Many CI fixes; switch from herumi to AntelopeIO/bn256

+++ src/l2_tracker/l2_tracker.h
@@ -78,7 +78,6 @@ class L2Tracker {
     void add_to_mempool(const event::StateChangeVariant& state_change);
 
   public:
-
     // Constructs an L2Tracker.  The optional `update_frequency` determines how frequently we poll
     // for updated chain state and logs: This defines how many requests we make to the L2 provider:
     // each (this period) we need to make one request to get the current block height, and then at
@@ -89,7 +88,7 @@ class L2Tracker {
     // is enough to have the last hour of history (for an L2 such as Arbitrum with its 0.25s block
     // time), plus a buffer so that pulse quorum nodes can properly recognize L2 events from the
     // past hour (for achieving pulse consensus).
-    size_t HIST_SIZE = 70min / 250ms;
+    uint64_t HIST_SIZE = 70min / 250ms;
 
     // How many blocks worth of logs we fetch at once.  Various providers impose various limits on
     // this based on the free/paid tier, and so there is no perfect default.  1000 blocks at once
